yml 파일은 반드시 /stock_web/config 폴더 내에 넣을 것
