+++ ticker_bread/modules/Visualize/Viz.py
@@ -1,0 +1,14 @@
+import pandas as pd
+import numpy as np
+import plotly.express as px
+
+class Viz:
+    def single_line_chart(self, data, x, y):
+        fig = px.line(data, x=x, y=y)
+
+        return fig
+    
+    def multi_line_chart(self, data, x, y, col):
+        fig = px.line(data, x=x, y=y, color=col)
+
+        return fig
